+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/services/QuestionService.java
@@ -8,7 +8,7 @@ import java.util.Set;
 
 public interface QuestionService extends CrudService<Question, Long> {
 
-    Set<Question> findAllByWordingEnLike(String searchPattern);
+    Set<Question> findAllByWordingLike(String searchPattern);
 
     Set<Question> findByKeyTerm(KeyTerm keyTerm);