+++ knowthenics-web/src/test/java/ru/dpopkov/knowthenics/controllers/QuestionControllerTest.java
@@ -130,13 +130,13 @@ class QuestionControllerTest {
 
     @Test
     void testProcessFindQuestionForm() throws Exception {
-        when(questionService.findAllByWordingEnLike(anyString())).thenReturn(questions);
+        when(questionService.findAllByWordingLike(anyString())).thenReturn(questions);
 
         mockMvc.perform(post("/questions/find"))
                 .andExpect(status().isOk())
                 .andExpect(model().attributeExists("questions"))
                 .andExpect(view().name("questions/list"));
-        verify(questionService).findAllByWordingEnLike(anyString());
+        verify(questionService).findAllByWordingLike(anyString());
     }
 
     @Test